moved all utility fns to utility.py to simplify package, and make way to delete update_step
